Update message signer
